fix(winding/wayland): bind poll count at native width
Linux defines poll's descriptor count as pointer-width nfds_t rather than a 32-bit integer. Small counts happened to work on common 64-bit calling conventions despite the nominal ABI mismatch.

Declare the count as usize and pass the corresponding bigint value at the call boundary.

+++ packages/winding/wayland/protocol.ts
@@ -8,7 +8,7 @@ export const libcSymbols = {
   mmap: { parameters: ["pointer", "usize", "i32", "i32", "i32", "i64"], result: "pointer" },
   munmap: { parameters: ["pointer", "usize"], result: "i32" },
   close: { parameters: ["i32"], result: "i32" },
-  poll: { parameters: ["buffer", "u32", "i32"], result: "i32" },
+  poll: { parameters: ["buffer", "usize", "i32"], result: "i32" },
 } as const satisfies Deno.ForeignLibraryInterface;
 
 export type LibcLibrary = Deno.DynamicLibrary<typeof libcSymbols>;
